Instant cast

+++ GW2EIGW2API/GW2API/GW2APISkill.cs
@@ -44,9 +44,9 @@ namespace GW2EIGW2API.GW2API
         [JsonProperty(PropertyName = "prev_chain")]
         public int PrevChain { get; internal set; }
         [JsonProperty(PropertyName = "transform_skills")]
-        public List<int> TransformSkills { get; internal set; }
+        public List<long> TransformSkills { get; internal set; }
         [JsonProperty(PropertyName = "bundle_skills")]
-        public List<int> BundleSkills { get; internal set; }
+        public List<long> BundleSkills { get; internal set; }
 
         [JsonProperty(PropertyName = "toolbelt_skill")]
         public int ToolbeltSkill { get; internal set; }
